requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.4
blis==1.3.3
Brlapi==0.8.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
cloudpathlib==0.23.0
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.13
dbus-python==1.2.18
debugpy==1.8.17
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
duplicity==0.8.21
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.122.0
fasteners==0.14.1
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
future==0.18.2
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.7
importlib-metadata==4.6.4
ipykernel==7.1.0
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keyring==23.5.0
kiwisolver==1.4.9
labelImg==1.8.6
language-selector==0.1
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==6.0.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
mongoengine==0.29.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
murmurhash==1.0.15
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
notebook_shim==0.2.4
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
openai==2.11.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.5
pdfminer.six==20251107
pexpect==4.8.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.12.4
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.5.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
reportlab==3.6.8
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
roboflow==1.2.11
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.15.3
scour==0.38.2
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.39.0
Send2Trash==1.8.3
six==1.16.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
systemd-python==234
terminado==0.18.1
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.232
ultralytics-thop==2.0.18
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.5.0
usb-creator==0.3.7
uvicorn==0.38.0
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
wrapt==2.0.1
wsproto==1.3.2
xdg==5
xkit==0.0.0
zipp==1.0.0
